Add adjustable scrollOffset to make see/click work with sticky headers
Fixes `see(foo)` when the page has a sticky header or other content.
Native `scrollIntoView()` will scroll the element to the top of the page,
which will be behind the header. The element is thus not visible.

The usage is not great because it's just a "global" `setOffset`
at the top of your spec. Suggestions welcome.